Moved Job overlap code out of analyzer
